% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.functions.R
\name{calculate.pcas}
\alias{calculate.pcas}
\title{Calculates principal components for every cluster}
\usage{
calculate.pcas(X, segmentation, number.clusters, max.subspace.dim,
  estimate.dimensions, prev.segmentation, prev.pcas)
}
\arguments{
\item{X}{A data matrix.}

\item{segmentation}{A vector, segmentation of variables into clusters.}

\item{number.clusters}{An integer, number of subspaces (clusters).}

\item{max.subspace.dim}{An integer, upper bound for allowed dimension of subspace.}

\item{estimate.dimensions}{A boolean, if TRUE subspaces dimensions are estimated using PESEL.}

\item{prev.pcas}{A list, subset of principal components for every cluster from previous iteration.}

\item{prev.segmenation}{A vector, segmentation of variables into clusters from previous iteration.}
}
\value{
A subset of principal components for every cluster.
}
\description{
For given segmentation this function estimates dimensionality of each cluster (or chooses fixed dimensionality)
and for each cluster calculates the number of principal components equal to the this dimensionality
}
\keyword{internal}
