% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.functions.R
\name{choose.cluster.BIC}
\alias{choose.cluster.BIC}
\title{Selects subspace closest to a given variable (according to BIC)}
\usage{
choose.cluster.BIC(variable, pcas, numberClusters, show.warnings = FALSE)
}
\arguments{
\item{variable}{variable variable to be assigned}

\item{pcas}{orthogonal basis for different subspaces}

\item{numberClusters}{number of subspaces (clusters)}

\item{show.warnings}{a boolean - if set to TRUE all warnings are displayed, default value is FALSE}
}
\value{
index number of subspace closest to variable
}
\description{
Selects subspace closest to a given variable (according to BIC)
}
\keyword{internal}
