% Generated by roxygen2 (4.0.1): do not edit by hand
\name{choose.cluster}
\alias{choose.cluster}
\title{Choose subspace closest to the given variable}
\usage{
choose.cluster(variable, pcas, numberClusters)
}
\arguments{
\item{varaible}{}

\item{pcas}{orthogonal basis for different subspaces}

\item{numberClusters}{number of subspaces (clusters)}
}
\value{
index number of subspace closest to variable
}
\description{
The most similar subspace is choosen based on R^2
}

