% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cluster.pca.BIC}
\alias{cluster.pca.BIC}
\title{BIC for subspace clustering}
\usage{
cluster.pca.BIC(X, segmentation, dims, numb.clusters, max.dim)
}
\arguments{
\item{X}{a matrix with only continuous variables}

\item{segmentation}{a vector, segmentation for which likelihood is computed. Clusters
numbers should be from range [1, numb.clusters]}

\item{dims}{a vector of integers, dimensions of subspaces. Number of principal components
that span each subspace.}

\item{numb.clusters}{an integer, number of clusters}

\item{max.dim}{an integer, upper bound for allowed dimension of subspace}
}
\value{
BIC value of BIC criterion
}
\description{
Computes the value of BIC criterion for given data set and partition.
In each cluster we assume that variables are spanned by few factors.
Considering maximum likelihood we get that those factors are in fact
principal components. Noise sigma can be computed jointly for all clusters (default),
seperately for each cluster or be specified as input.
}
\keyword{internal}

