% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dataSIMULATION}
\alias{dataSIMULATION}
\title{Simulating subspace clustering data}
\usage{
dataSIMULATION(n = 100, SNR, K = 20, numbPoints = 50, max_dim = 1)
}
\arguments{
\item{n}{number of observations}

\item{SNR}{signal to noise ratio}

\item{K}{number of subspaces}

\item{numbPoints}{number of variables in each subspace}

\item{max.dim}{maximum dimension of subspace}
}
\value{
a list consisting of
\item{X}{generated data}
\item{signals}{data without noise}
\item{dims}{dimensions of subspaces}
\item{s}{true clustering}
}
\description{
Generating random data for subspace clustering simulation
}

