% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estim.ncp.R
\name{estim.ncp}
\alias{estim.ncp}
\title{Estimating number of relevant (non noise) Principal Components in PCA}
\usage{
estim.ncp(X, ncp.min = 1, ncp.max = 10, scale = TRUE, verbose = FALSE)
}
\arguments{
\item{X}{a data frame or a matrix contatining only continuous variables}

\item{ncp.min}{minimal number of principal components, for all the possible number of
PCs between ncp.min and ncp.max criterion is computed}

\item{ncp.max}{maximal number of principal components, if greater than dimensions of X,
min(ncol(X), nrow(X))-1 is used, for all the possible number of
PCs between ncp.min and ncp.max criterion is computed}

\item{scale}{a boolean, if TRUE (default value) then data is scaled before applying
criterion}

\item{verbose}{a boolean, if TRUE plot with BIC values for different
numbers of components is produced (default value is FALSE)}
}
\value{
number of components
}
\description{
For a given numeric data set, estimates number of PCs according to
penalized likelihood criterion
}
\examples{
\dontrun{
library(MetabolAnalyze)
data(UrineSpectra)
estim.ncp(UrineSpectra[[1]], verbose=TRUE)}
}

