% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSigma}
\alias{getSigma}
\title{Computing joint sigma for all clusters

Computes unbiased noise estimator under assumption that
all subspaces are of the same dimension}
\usage{
getSigma(X, segmentation, max.dim, n, p, numb.clusters)
}
\arguments{
\item{X}{data}

\item{segmentation}{variable segmentation}

\item{max.dim}{maximal subspace dimension}

\item{n}{number of rows}

\item{p}{number of variables}

\item{numb.clusters}{total number of clusters}
}
\value{
unbiased noise estimator
}
\description{
Computing joint sigma for all clusters

Computes unbiased noise estimator under assumption that
all subspaces are of the same dimension
}
\author{
Piotr Sobczyk
}
\keyword{internal}

