% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integration}
\alias{integration}
\title{Computes integration and acontamination of each cluster in given partition}
\usage{
integration(group, true_group)
}
\arguments{
\item{group}{a vector, first partition}

\item{true_group}{a vector, second (reference) partition}
}
\value{
(integration, acontamination)
}
\description{
Computes integration and acontamination of each cluster in given partition
}
\examples{
\donttest{
sim.data <- data.simulation(n = 20, SNR = 1, K = 2, numb.vars = 50, max.dim = 2)
true_segmentation <- rep(1:2, each=50)
mlcc.fit <- mlcc.reps(sim.data$X, numb.clusters = 2, max.dim = 2)
integration(true_segmentation, mlcc.fit$segmentation)
}
}
\references{
{M. Sołtys. Metody analizy skupień. Master’s thesis, Wrocław University of Technology, 2010}
}
