% Generated by roxygen2 (4.0.1): do not edit by hand
\name{missclassify.heuristic}
\alias{missclassify.heuristic}
\title{Compute missclasification rate for subspace clustering}
\usage{
missclassify.heuristic(group, N, n)
}
\arguments{
\item{group}{proposed clustering}

\item{N}{number of elements in each cluster}

\item{n}{number of clusters}
}
\value{
mis misclassification rate
}
\description{
As getting exact value requires checking all permutations a heuristic approach is proposed
It is assumed that there are n cluster each of N elements and that they are sorted.
}

