% Generated by roxygen2 (4.0.1): do not edit by hand
\name{missclassify.heuristic2}
\alias{missclassify.heuristic2}
\title{Compute missclasification rate for subspace clustering}
\usage{
missclassify.heuristic2(group, true_group, N, n)
}
\arguments{
\item{group}{proposed clustering}

\item{true_group}{true clustering}

\item{N}{maximal assumed number of elements in cluster}

\item{n}{number of clusters}
}
\value{
mis misclassification rate
}
\description{
As getting exact value requires checking all permutations a heuristic approach is proposed
It is assumed that there are n cluster each of N elements and that they are sorted.
}

