% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MPCV.BIC}
\alias{MPCV.BIC}
\title{Subspace clustering with automatic estimation of number of clusters
and their dimension.}
\usage{
MPCV.BIC(X, numb.Clusters = 1:10, numb.runs = 20, stop.criterion = 1,
  max.iter = 20, max.dim = 1, scale = T, numbCores = 1, greedy = TRUE,
  estimateDimensions = T)
}
\arguments{
\item{X}{a data frame or a matrix with only continuous variables}

\item{numb.Clusters}{a vector, numbers of clusters to be checked}

\item{numb.runs}{an integer, number of runs of MLCC}

\item{stop.criterion}{an integer, indicating how many changes in partitions triggers stopping the MLCC algorithm}

\item{max.iter}{an integer, maximum number of iterations of MLCC algorithm}

\item{max.dim}{an integer, maximum dimension of subspaces to be considered}

\item{scale}{a boolean, if TRUE (value set by default) then data are scaled to unit variance}

\item{numbCores}{an integer, number of cores to be used, by default all cores are used}

\item{greedy}{a boolean, if TRUE (value set by default) the clusters are estimated in a greedy way}

\item{estimateDimensions}{a boolean, if TRUE (value set by default) subspaces dimensions are estimated}
}
\value{
An object of class mlcc.fit consisting of
\item{segmentation}{a vector containing the partition of the variables}
\item{BIC}{double, value of \code{\link{myBIC}} criterion}
\item{subspacesDimensions}{a list containing dimensions of the subspaces}
\item{nClusters}{an integer, estimated number of clusters}
}
\description{
Estimate the number of clusters according to the BIC. Basic k-means based
Multiple Latent Components Clustering (MLCC) algorithm is run a
specified number of times (numb.runs) for each number of clusters in numb.Clusters
}
\examples{
\donttest{
data <- dataSIMULATION(n=100, SNR=1, K=5, numbVars=30, max.dim=2)
MPCV.BIC(data$X, numb.Clusters=1:10, numb.runs=20)}
}

