% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MPCV}
\alias{MPCV}
\title{Subspace clustering based on multiple principal components}
\usage{
MPCV(X, numberClusters = 2, stop.criterion = 1, max.iter = 40,
  maxSubspaceDim = 4, initial.segmentation = NULL)
}
\arguments{
\item{X}{data}

\item{numberClusters}{number of clusters to be used}

\item{stop.criterion}{}

\item{max.iter}{}

\item{maxSubspaceDim}{}

\item{initial.segmentation}{}
}
\value{
a list consisting of
\item{segmentation}{of points to clusters}
\item{pcas}{basis of each cluster}
}
\description{
Performs subspace clustering
}

