% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MPCV.reps}
\alias{MPCV.reps}
\title{Subspace clustering based on multiple principal components}
\usage{
MPCV.reps(X, numb.Clusters = 2, numb.runs = 20, stop.criterion = 1,
  max.iter = 20, initial.segmentation = NULL, max.dim = 1,
  method = c("likelihood", "singular", "residual"), scale = T,
  numbCores = 1)
}
\arguments{
\item{X}{data}

\item{numb.Clusters}{number of clusters to be used}

\item{numb.runs}{number of runs of MLCC}

\item{stop.criterion}{how many changes in partitions triggers stopping the algorithm}

\item{max.iter}{maxium number of iteratations}

\item{initial.segmentation}{initial segmentation of variable to clusters}

\item{max.dim}{maximum considered dimension of subspaces}

\item{method}{method to be used to determine best run. Possible values are "likelihood", "singular", "residual"}

\item{scale}{Should data be scaled?}

\item{numbCores}{Number of cores to be used}
}
\value{
a list consisting of
\item{segmentation}{of points to clusters}
\item{BIC}{Value of \code{\link{myBIC}} criterion}
}
\description{
Performs MPCC multiple times and chooses the best run based on residual variance
}

