% Generated by roxygen2 (4.0.2): do not edit by hand
\name{myBIC_one_cluster}
\alias{myBIC_one_cluster}
\title{BIC for subspace clustering}
\usage{
myBIC_one_cluster(X, max.dim, numb.clusters, sigma = NULL)
}
\arguments{
\item{X}{data}

\item{max.dim}{maximum dimension of subspace}

\item{numb.clusters}{number of clusters}

\item{sigma}{(optional) pre-computed value of sigma}
}
\value{
BIC value of BIC criterion
}
\description{
Computes the value of BIC criterion
}

