% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pca.Laplace}
\alias{pca.Laplace}
\title{Laplace evidence for PCA, as given by Minka}
\usage{
pca.Laplace(X, k, alfa = 1)
}
\arguments{
\item{X}{a matrix with only continuous variables}

\item{k}{number of principal components fitted}
}
\value{
L value of Laplace approximation
}
\description{
Computes the value of Laplace approximation for given data set and
number of factors.
}
\references{
Automatic choice of dimensionality for PCA, Thomas P. Minka
}
\keyword{internal}

