% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.functions.R
\name{pca.new.BIC}
\alias{pca.new.BIC}
\title{Penalized likelihood for PCA}
\usage{
pca.new.BIC(X, k, show.warnings = FALSE)
}
\arguments{
\item{X}{a matrix with only continuous variables}

\item{k}{number of principal components fitted}

\item{show.warnings}{a boolean - if set to TRUE all warnings are displayed, default value is FALSE}
}
\value{
BIC value of BIC criterion
}
\description{
Computes the value of BIC-like criterion for given data set and 
number of factors. Assumes that number of variables is large
compared to number of observations
}
\keyword{internal}
