% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.functions.R
\name{show.clusters}
\alias{show.clusters}
\title{Print clusters obtained from MLCC}
\usage{
show.clusters(data, segmentation)
}
\arguments{
\item{data}{The original data set.}

\item{segmentation}{A vector, segmentation of variables into clusters.}
}
\description{
Print clusters obtained from MLCC
}
