% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.functions.R
\name{calculate.distance.kmeanspp}
\alias{calculate.distance.kmeanspp}
\title{Calculates pseudo-distance from given variable to one dimensional clusters by BIC (initialization of kmeans++)}
\usage{
calculate.distance.kmeanspp(variable, pcas, numberClusters)
}
\arguments{
\item{variable}{variable for which the pseudo-distance is calculated}

\item{pcas}{orthogonal basis for different subspaces}

\item{numberClusters}{number of subspaces (clusters)}
}
\value{
minimal distance to a subspace
}
\description{
Calculates pseudo-distance from given variable to one dimensional clusters by BIC (initialization of kmeans++)
}
\keyword{internal}
