% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.functions.R
\name{calculate.pcas}
\alias{calculate.pcas}
\title{Calculates principal components for every cluster and chooses the dimesionality using PESEL criterion.}
\usage{
calculate.pcas(X, segmentation, number.clusters, max.subspace.dim,
  estimate.dimensions)
}
\arguments{
\item{X}{a data matrix}

\item{segmentation}{a vector, segmentation of variables into clusters}

\item{number.clusters}{number of subspaces (clusters)}

\item{max.subspace.dim}{an integer, upper bound for allowed dimension of subspace}

\item{estimate.dimensions}{a boolean, if TRUE subspaces dimensions are estimated}
}
\value{
A subset of principal components for every cluster
}
\description{
Calculates principal components for every cluster and chooses the dimesionality using PESEL criterion.
}
\keyword{internal}
