% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.functions.R
\name{choose.cluster.BIC}
\alias{choose.cluster.BIC}
\title{Assigns variables to subspaces (according to BIC)}
\usage{
choose.cluster.BIC(variable, pcas, numberClusters)
}
\arguments{
\item{variable}{variable variable to be assigned}

\item{pcas}{orthogonal basis for different subspaces}

\item{numberClusters}{number of subspaces (clusters)}
}
\value{
index number of subspace closest to variable
}
\description{
Selects subspace closest to given variable (according to BIC)
}
\keyword{internal}
