% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.functions.R
\name{cluster.pca.BIC}
\alias{cluster.pca.BIC}
\title{mBIC for subspace clustering}
\usage{
cluster.pca.BIC(X, segmentation, dims, numb.clusters, max.dim,
  flat.prior = FALSE)
}
\arguments{
\item{X}{a matrix with only continuous variables}

\item{segmentation}{a vector, segmentation for which likelihood is computed. Clusters
numbers should be from range [1, numb.clusters]}

\item{dims}{a vector of integers, dimensions of subspaces. Number of principal components
(chosen by PESEL criterion) that span each subspace.}

\item{numb.clusters}{an integer, number of clusters}

\item{max.dim}{an integer, upper bound for allowed dimension of subspace}

\item{flat.prior}{boolean, if TRUE (default is FALSE) then flat prior on models is used}
}
\value{
value of mBIC
}
\description{
Computes the value of modified Bayesian Information Criterion (mBIC)
for given data set and partition. In each cluster we assume that 
variables are spanned by few factors.
Considering maximum likelihood we get that those factors are in fact
principal components. Additionally, it uses by default an informative 
prior distribution on models.
}
\keyword{internal}
