% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estim.ncp.R
\name{estim.ncp}
\alias{estim.ncp}
\title{Estimating number of important Principal Components in PCA}
\usage{
estim.ncp(X, ncp.min = 1, ncp.max = 10, method = c("penLik", "RajanPen",
  "minkaBIC", "minkaLaplace"), scale = TRUE, verbose = FALSE)
}
\arguments{
\item{X}{a data frame or a matrix with only continuous variables}

\item{ncp.min}{minimal number of principal components}

\item{ncp.max}{maximal number of principal components}

\item{method}{a crtierion to be used}

\item{scale}{a boolean, if TRUE (default value) then data is scaled}

\item{verbose}{a boolean, if TRUE plot with BIC values for different
numbers of components is produced}
}
\value{
Number of components
}
\description{
For a given data set, estimates numberof PCs according to
one of four different penalized likelihood criterions
}
\references{
, P. Sobczyk, M. Bogdan, J. Josse,
Bayesian dimensionality reduction with PCA using partially integrated penalized likelihood
}

