% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integration}
\alias{integration}
\title{Computes integration of each cluster in given partition}
\usage{
integration(group, true_group)
}
\arguments{
\item{group}{a vector, first partition}

\item{true_group}{a vector, second (reference) partition}
}
\value{
integration
}
\description{
Computes integration of each cluster in given partition
}
\references{
{M. Sołtys. Metody analizy skupień. Master’s thesis, Wrocław University of Technology, 2010}
}
