% Generated by roxygen2 (4.0.2): do not edit by hand
\name{missclassify.heuristic2}
\alias{missclassify.heuristic2}
\title{Computes missclasification rate.}
\usage{
missclassify.heuristic2(group, true_group, N, n)
}
\arguments{
\item{group}{a vector, first partiton}

\item{true_group}{a vector, second (reference) partition}

\item{N}{an integer, maximal number of elements in one class}

\item{n}{an integer, number of classes}
}
\value{
misclassification rate
}
\description{
As getting exact value of missclasification requires checking all permutations
and is therefore intrackable even for modest number of classes, a heuristic approach is proposed.
It is assumed that there are n classes of maximum N elements.
Additional requirement is that classes identifiers are from range [1, n]
}
\examples{
partition1 <- sample(10, 300, replace=T)
partition2 <- sample(10, 300, replace=T)
missclassify.heuristic2(partition1, partition1, max(table(partition1)), 10)
missclassify.heuristic2(partition1, partition2, max(table(partition2)), 10)
}

