% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mlcc.bic}
\alias{mlcc.bic}
\title{Subspace clustering with automatic estimation of number of clusters
and their dimension}
\usage{
mlcc.bic(X, numb.clusters = 1:10, numb.runs = 20, stop.criterion = 1,
  max.iter = 20, max.dim = 4, scale = TRUE, numb.cores = NULL,
  greedy = TRUE, estimate.dimensions = TRUE, verbose = FALSE)
}
\arguments{
\item{X}{a data frame or a matrix with only continuous variables}

\item{numb.clusters}{a vector, numbers of clusters to be checked}

\item{numb.runs}{an integer, number of runs of \code{\link{mlcc.kmeans}}}

\item{stop.criterion}{an integer, if an iteration of \code{\link{mlcc.kmeans}} algorithm
makes less changes in partitions than \code{stop.criterion},
\code{\link{mlcc.kmeans}} stops.}

\item{max.iter}{an integer, maximum number of iterations of \code{\link{mlcc.kmeans}} algorithm}

\item{max.dim}{an integer, if estimate.dimensions is FALSE then max.dim is dimension of each subspace.
If estimate.dimensions is TRUE then subspaces dimensions are estimated from the range [1, max.dim]}

\item{scale}{a boolean, if TRUE (value set by default) then variables in
dataset are scaled to zero mean and unit variance}

\item{numb.cores}{an integer, number of cores to be used, by default all cores are used}

\item{greedy}{a boolean, if TRUE (value set by default) the clusters are estimated in a greedy way}

\item{estimate.dimensions}{a boolean, if TRUE (value set by default) subspaces dimensions are estimated}

\item{verbose}{a boolean, if TRUE plot with BIC values for different
numbers of clusters is produced and values of BIC, computed
for every number of clusters and subspaces dimensions, are printed
(value set by default is FALSE)}
}
\value{
An object of class mlcc.fit consisting of
\item{segmentation}{a vector containing the partition of the variables}
\item{BIC}{numeric, value of \code{\link{cluster.BIC}} criterion}
\item{subspacesDimensions}{a list containing dimensions of the subspaces}
\item{nClusters}{an integer, estimated number of clusters}
\item{factors}{a list of matrices, basis for each subspace}
\item{all.fit}{a list of segmentation, BIC, subspaces dimension for all numbers of clusters considered for an estimated subspace dimensions}
\item{all.fit.dims}{a list of lists of segmentation, BIC, subspaces dimension for all numbers of clusters and subspaces dimensions considered}
}
\description{
Estimate the number of clusters according to the BIC. Basic k-means based
Multiple Latent Components Clustering (MLCC) algorithm (\code{\link{mlcc.kmeans}}) is run a
given number of times (\emph{numb.runs}) for each number of clusters in \emph{numb.clusters}.
The best partition is choosen with BIC (see \code{\link{mlcc.reps} function})
}
\examples{
\donttest{
sim.data <- data.simulation(n = 100, SNR = 1, K = 5, numb.vars = 30, max.dim = 2)
mlcc.bic(sim.data$X, numb.clusters = 1:10, numb.runs = 20, verbose=TRUE)
}
}

