% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlcc.kmeans.R
\name{mlcc.kmeans}
\alias{mlcc.kmeans}
\title{Multiple Latent Components Clustering - kmeans algorithm}
\usage{
mlcc.kmeans(X, number.clusters = 2, stop.criterion = 1, max.iter = 40,
  max.subspace.dim = 4, initial.segmentation = NULL,
  estimate.dimensions = TRUE, mode = "random", show.warnings = FALSE)
}
\arguments{
\item{X}{a matrix with only continuous variables}

\item{number.clusters}{an integer, number of clusters to be used}

\item{stop.criterion}{an integer indicating how many changes in partitions triggers stopping the algorithm}

\item{max.iter}{an integer, maximum number of iterations of k-means}

\item{max.subspace.dim}{an integer, maximum dimension of subspaces}

\item{initial.segmentation}{a vector, initial segmentation of variables to clusters}

\item{estimate.dimensions}{a boolean, if TRUE (value set by default) subspaces dimensions are estimated}

\item{mode}{a string, possible values : random, kmeans++, sPCA, determines the intialization mode of the algorithm}

\item{show.warnings}{a boolean - if set to TRUE all warnings are displayed, default value is FALSE}
}
\value{
A list consisting of:
\item{segmentation}{a vector containing the partition of the variables}
\item{pcas}{a list of matrices, basis vectors for each cluster (subspace)}
}
\description{
Performs k-means based subspace clustering. Center of each cluster is some number 
of principal components. Similarity measure is calculated using BIC.
}
\examples{
\donttest{
sim.data <- data.simulation(n = 50, SNR = 1, K = 5, numb.vars = 50, max.dim = 3)
mlcc.kmeans(sim.data$X, number.clusters = 5, max.iter = 20, max.subspace.dim = 3, mode = "kmeans++")
}
}
