% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlcc.reps.R
\name{mlcc.reps}
\alias{mlcc.reps}
\title{Subspace clustering assuming that the number of clusters is known}
\usage{
mlcc.reps(X, numb.clusters = 2, numb.runs = 30, stop.criterion = 1,
  max.iter = 40, initial.segmentations = NULL, max.dim = 4,
  scale = TRUE, numb.cores = NULL, estimate.dimensions = TRUE,
  flat.prior = FALSE, show.warnings = FALSE, deterministic = FALSE)
}
\arguments{
\item{X}{a data frame or a matrix with only continuous variables}

\item{numb.clusters}{an integer, number of cluster}

\item{numb.runs}{an integer, number of runs of \code{\link{mlcc.kmeans} algorithm} with random initialization}

\item{stop.criterion}{an integer, if an iteration of \code{\link{mlcc.kmeans}} algorithm 
makes less changes in partitions than \code{stop.criterion}, 
\code{\link{mlcc.kmeans}} stops.}

\item{max.iter}{an integer, maximum number of iterations of \code{\link{mlcc.kmeans}} algorithm}

\item{initial.segmentations}{a list of vectors, segmentations that user wants to be 
used as an initial segmentation in \code{\link{mlcc.kmeans}} algorithm}

\item{max.dim}{an integer, dimension of subspaces (all are assumed to be equal)}

\item{scale}{a boolean, if TRUE (value set by default) then variables in 
dataset are scaled to zero mean and unit variance}

\item{numb.cores}{an integer, number of cores to be used, by default all cores are used}

\item{estimate.dimensions}{a boolean, if TRUE (value set by default) subspaces dimensions are estimated}

\item{flat.prior}{a boolean, if TRUE then, instead of a prior that takes into account
number of models for a given number of clusters, flat prior is used}

\item{show.warnings}{a boolean - if set to TRUE all warnings are displayed, default value is FALSE}

\item{deterministic}{a boolean - if set to TRUE the results of the method will be reproductible
(run with the same parameters will return the same result), default value is FALSE}
}
\value{
A list consisting of
\item{segmentation}{a vector containing the partition of the variables}
\item{BIC}{a numeric, value of \code{\link{cluster.pca.BIC}} criterion}
\item{basis}{a list of matrices, the basis vectors for subspaces}
}
\description{
For a fixed number of cluster function returns the best partition and basis for each subspace.
}
\details{
In more detail, an algorithm \code{\link{mlcc.kmeans}} 
is run a \emph{numb.runs} of times with random or custom initializations. 
The best partition is selected according to the BIC.
}
\examples{
\donttest{
sim.data <- data.simulation(n = 50, SNR = 1, K = 5, numb.vars = 50, max.dim = 3)
mlcc.reps(sim.data$X, numb.clusters = 5, numb.runs = 20, max.dim = 4)
}
}
