% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MPCV.BIC}
\alias{MPCV.BIC}
\title{Subspace clustering based on multiple principal components}
\usage{
MPCV.BIC(X, numb.Clusters = 1:10, numb.runs = 20, stop.criterion = 1,
  max.iter = 20, max.dim = 1, method = c("likelihood", "singular",
  "residual"), scale = T, numbCores = 1, greedy = TRUE,
  estimateDimensions = T)
}
\arguments{
\item{X}{data}

\item{numb.Clusters}{clusters numbers to check}

\item{numb.runs}{number of runs of MLCC}

\item{stop.criterion}{how many changes in partitions triggers stopping the algorithm}

\item{max.iter}{maxium number of iteratations}

\item{max.dim}{maximum considered dimension of subspaces}

\item{method}{method to be used to determine best run. Possible values are "likelihood", "singular", "residual"}

\item{scale}{Should data be scaled?}

\item{numbCores}{Number of cores to be used}

\item{greedy}{Should number of clusters be estimated in a greedy way}

\item{estimateDimensions}{Should subspaces dimensions be estimated as well?}
}
\value{
a list consisting of
\item{segmentation}{of points to clusters}
\item{BIC}{Value of \code{\link{myBIC}} criterion}
}
\description{
Performs MPCC multiple times and chooses the best run based on residual variance
}

