% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MPCV.reps}
\alias{MPCV.reps}
\title{Subspace clustering assuming that the number of clusters is known}
\usage{
MPCV.reps(X, numb.Clusters = 2, numb.runs = 20, stop.criterion = 1,
  max.iter = 20, initial.segmentations = NULL, max.dim = 2, scale = T,
  numbCores = NULL)
}
\arguments{
\item{X}{a data frame or a matrix with only continuous variables}

\item{numb.Clusters}{an integer, number of cluster}

\item{numb.runs}{an integer, number of runs of \code{\link{MPCV} algorithm} with random initialization}

\item{stop.criterion}{an integer, indicating how many changes in partitions triggers stopping the MLCC algorithm}

\item{max.iter}{an integer, maximum number of iterations of MLCC algorithm}

\item{initial.segmentations}{a list of vectors, segmentations that user wants to be used as an initial segmentation in MLCC algorithm}

\item{max.dim}{an integer, maximum dimension of subspaces to be considered}

\item{scale}{a boolean, if TRUE (value set by default) then data are scaled to unit variance}

\item{numbCores}{an integer, number of cores to be used, by default all cores are used}
}
\value{
A list consisting of
\item{segmentation}{a vector containing the partition of the variables}
\item{BIC}{double, value of \code{\link{myBIC}} criterion}
}
\description{
For a fixed number of cluster and fixed number of components per cluster,
select the best partition according to the BIC.
}
\examples{
\donttest{
data <- dataSIMULATION(n=100, SNR=1, K=5, numbVars=30, max.dim=2)
MPCV.reps(data$X, numb.Clusters=5, numb.runs=20)}
}

