% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MPCV.reps}
\alias{MPCV.reps}
\title{Subspace clustering based on multiple principal components}
\usage{
MPCV.reps(X, numb.Clusters = 2, numb.runs = 20, stop.criterion = 1,
  max.iter = 20, max.dim = 1, method = "h", scale = T)
}
\arguments{
\item{X}{data}

\item{numberClusters}{number of clusters to be used}

\item{numb.runs}{number of runs of MPCV}

\item{stop.criterion}{}

\item{max.iter}{}

\item{max.dim}{}

\item{method}{method to be used to determine best run. Possible values are 'h' and 'r'}

\item{scale}{Should data be scaled?}
}
\value{
a list consisting of
\item{segmentation}{of points to clusters}
\item{BIC}{Value of BIC criterion}
}
\description{
Performs MPCC multiple times and chooses the best run based on residual variance
}

