% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pca.new.BIC}
\alias{pca.new.BIC}
\title{Version of BIC for PCA}
\usage{
pca.new.BIC(X, k)
}
\arguments{
\item{X}{a matrix with only continuous variables}

\item{k}{number of principal components fitted}
}
\value{
BIC value of BIC criterion
}
\description{
Computes the value of BIC criterion for given data set and
number of factors.
}
\references{
Automatic choice of dimensionality for PCA, Thomas P. Minka
}
\keyword{internal}

