\name{VarClust-package}
\alias{VarClust-package}
\alias{VarClust}
\docType{package}
\title{
Variable Clustering
}
\description{
Variable Clustering based on k-means algorithm. In each step cluster center is a number of PCA components
  of points in that cluster. Than distance is defined as R^2 in least-squares.
}
\details{
\tabular{ll}{
Package: \tab VarClust\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-16\cr
License: \tab What license is it under?\cr
}
Use function MPCV.
}
\author{
Author: Piotr Sobczyk
Maintainer: Piotr Sobczyk <piotr.sobczyk@pwr.edu.pl>

}
\references{
Ulrike von Luxburg 2007 \emph{A Tutorial on Spectral Clustering} 

Brian McWilliams, Giovanni Montana 2012 \emph{Subspace clustering of high-dimensional data: a predictive approach} 

Marie Chavent, Benoit Liquet, Vanessa Kuentz-Simonet, Jerome Saracco 2013 \emph{ClustOfVar: An R Package for the Clustering of Variables} 
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
