\name{VarClust-package}
\alias{VarClust-package}
\alias{VarClust}
\docType{package}
\title{
Variable Clustering
}
\description{
Variable Clustering based on k-means algorithm. In each step cluster center is a number of PCA components
  of points in that cluster. Than distance is defined as R^2 in least-squares.
}
\details{
\tabular{ll}{
Package: \tab VarClust\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-16\cr
License: \tab What license is it under?\cr
}
Use function MPCV.
}
\author{
Author: Piotr Sobczyk
Maintainer: Piotr Sobczyk <piotr.sobczyk@pwr.edu.pl>

}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
