% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{VarClust}
\alias{VarClust}
\alias{VarClust-package}
\title{Variable Clustering}
\description{
Variable Clustering based on k-means algorithm. In each step cluster center is a number of PCA components
of points in that cluster. Than distance is defined as R^2 in least-squares.
}
\details{
The main function of package \pkg{VarClust} is
\code{\link{MPCV.BIC}} which allows you to cluster data
with unknown number of clusters. If you have know
the number of clusters use function \code{\link{MPCV.reps}}

Version: 0.9
}
\examples{
\dontrun{MPCV.reps(data)}
}
\author{
Piotr Sobczyk

Maintainer: Piotr Sobczyk \email{Piotr.Sobczyk@pwr.edu.pl}
}
\references{
Ulrike von Luxburg 2007 \emph{A Tutorial on Spectral Clustering}

Marie Chavent, Benoit Liquet, Vanessa Kuentz-Simonet, Jerome Saracco 2013 \emph{ClustOfVar: An R Package for the Clustering of Variables}
}

